+++ src/Layer/PaintNodePrivate.hpp
@@ -428,7 +428,6 @@ public:
         break;
       }
     }
-
     {
       if (filled)
       {
